requirements.txt
ollama
chromadb
fastapi
uvicorn
python-multipart
python-dotenv
